requirements.txt
timm==0.9.2
torch==1.12.1+cu113
torchaudio==0.12.1+cu113
torchcam==0.3.2
torchstat==0.0.7
torchvision==0.13.1+cu113
tornado==6.3.3
tqdm==4.66.4
albumentations==1.3.1
efficientnet-pytorch==0.7.1
grad-cam==1.5.0
graphviz==0.20.3
grpcio==1.60.0
h5py==3.10.0
huggingface-hub==0.24.0
imageio==2.33.0
imgviz==1.7.5
ipykernel==6.27.1
json5==0.9.14
keras==2.11.0
keras-cv-attention-models==1.3.9
labelme==5.5.0
matplotlib==3.8.2
matplotlib-inline==0.1.6
MedPy==0.4.0
monai==1.3.0
multiprocess==0.70.16
numba==0.60.0
numpy==1.26.2
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
opt-einsum==3.3.0
pandas==2.1.3
pandas-flavor==0.6.0
Pillow==10.1.0
pretrainedmodels==0.7.4
python-json-logger==2.0.7
pytorch-fid==0.3.0
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.12.0
seaborn==0.13.1
statsmodels==0.14.1
tensorboard==2.11.2
thop==0.1.1.post2209072238


